+++ palette/src/lib.rs
@@ -241,6 +241,7 @@ pub use luma::{GammaLuma, GammaLumaa, LinLuma, LinLumaa, SrgbLuma, SrgbLumaa};
 pub use luv::{Luv, Luva};
 pub use okhsl::{Okhsl, Okhsla};
 pub use okhsv::{Okhsv, Okhsva};
+pub use okhwb::{Okhwb, Okhwba};
 pub use oklab::{Oklab, Oklaba};
 pub use oklch::{Oklch, Oklcha};
 pub use rgb::{GammaSrgb, GammaSrgba, LinSrgb, LinSrgba, Srgb, Srgba};
@@ -442,8 +443,10 @@ pub mod luma;
 mod luv;
 mod luv_bounds;
 pub mod num;
+mod ok_utils;
 mod okhsl;
 mod okhsv;
+mod okhwb;
 mod oklab;
 mod oklch;
 mod relative_contrast;
