+++ palette_derive/src/lib.rs
@@ -39,7 +39,7 @@ mod util;
 
 const COLOR_TYPES: &[&str] = &[
     "Rgb", "Luma", "Hsl", "Hsluv", "Hsv", "Hwb", "Lab", "Lch", "Lchuv", "Luv", "Oklab", "Oklch",
-    "Xyz", "Yxy",
+    "Okhwb", "Okhsl", "Okhsv", "Xyz", "Yxy",
 ];
 
 const PREFERRED_CONVERSION_SOURCE: &[(&str, &str)] = &[
@@ -55,6 +55,9 @@ const PREFERRED_CONVERSION_SOURCE: &[(&str, &str)] = &[
     ("Luv", "Xyz"),
     ("Oklab", "Xyz"),
     ("Oklch", "Oklab"),
+    ("Okhsl", "Oklab"),
+    ("Okhsv", "Oklab"),
+    ("Okhwb", "Okhsv"),
     ("Yxy", "Xyz"),
 ];
